+++ demo-formatter/go/cucumber_demo_formatter.go
@@ -7,7 +7,7 @@ package cucumber_demo_formatter
 
 import (
 	"fmt"
-	"github.com/cucumber/messages-go/v11"
+	"github.com/cucumber/messages-go/v12"
 	gio "github.com/gogo/protobuf/io"
 	"io"
 )