+++ react/javascript/src/components/gherkin/Description.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import marked from 'marked'
 import sanitizeHtml from 'sanitize-html'
+import HighLight from '../app/HighLight'
 
 interface IProps {
   description: string
@@ -9,7 +10,8 @@ interface IProps {
 const Description: React.FunctionComponent<IProps> = ({ description }) => {
   const html = marked(description)
   const sanitizedHtml = sanitizeHtml(html)
-  return <div dangerouslySetInnerHTML={{ __html: sanitizedHtml }} />
+
+  return <HighLight text={sanitizedHtml} htmlText={true} />
 }
 
 export default Description