+++ react/javascript/test/search/StepSearchTest.ts
@@ -11,9 +11,13 @@ describe('StepSearch', () => {
     stepSearch = new StepSearch()
 
     steps = [
-      makeStep('Given', 'a passed step'),
+      makeStep('Given', 'a passed step', 'There is a docstring here'),
       makeStep('When', 'another passed step'),
-      makeStep('Then', 'a failed step'),
+      makeStep('Then', 'a failed step', '', [
+        ['name', 'value'],
+        ['errorType', 'NullPointerException'],
+        ['message', 'Something really bad hapenned here'],
+      ]),
     ]
 
     for (const step of steps) {
@@ -49,6 +53,16 @@ describe('StepSearch', () => {
       // By default, ElasticLurn exclude some words from indexing/searching,
       // amongst them are 'Then' and 'When'.
       // See: http://elasticlunr.com/docs/stop_word_filter.js.html#resetStopWords
+    })
+
+    it('returns step which DocString matches the query', () => {
+      const searchResults = stepSearch.search('docstring')
+      assert.deepStrictEqual(searchResults, [steps[0]])
+    })
+
+    it('returns step which datatable matches the query', () => {
+      const searchResults = stepSearch.search('NullPointerException')
+      assert.deepStrictEqual(searchResults, [steps[2]])
     })
   })
 })