+++ tag-expressions/java/src/main/java/io/cucumber/tagexpressions/TagExpressionException.java
@@ -1,7 +1,7 @@
 package io.cucumber.tagexpressions;
 
-public class TagExpressionException extends RuntimeException {
-    public TagExpressionException(String message, Object... args) {
-        super(String.format(message, args));
+public final class TagExpressionException extends RuntimeException {
+    TagExpressionException(String message, Object... args) {
+		super(String.format(message, args));
     }
 }