+++ tag-expressions/java/src/test/java/io/cucumber/tagexpressions/TagExpressionParserParameterizedTest.java
@@ -9,8 +9,7 @@ import java.util.stream.Stream;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-public class TagExpressionParserParameterizedTest {
-    private final TagExpressionParser parser = new TagExpressionParser();
+class TagExpressionParserParameterizedTest {
 
     static Stream<Arguments> data() {
         return Stream.of(
@@ -25,8 +24,8 @@ public class TagExpressionParserParameterizedTest {
 
     @ParameterizedTest
     @MethodSource("data")
-    public void parser_expression(final String infix, final String expected) {
-        Expression expr = parser.parse(infix);
+    public void parser_expression(String infix, String expected) {
+        Expression expr = TagExpressionParser.parse(infix);
         assertEquals(expected, expr.toString());
     }