+++ tag-expressions/java/src/test/java/io/cucumber/tagexpressions/TagExpressionParserSyntaxErrorTest.java
@@ -7,35 +7,31 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-public class TagExpressionParserSyntaxErrorTest {
-
-    private final TagExpressionParser parser = new TagExpressionParser();
+class TagExpressionParserSyntaxErrorTest {
 
     static Stream<Arguments> data() {
         return Stream.of(
-                arguments("@a @b or", "Syntax error. Expected operator"),
-                arguments("@a and (@b not)", "Syntax error. Expected operator"),
-                arguments("@a and (@b @c) or", "Syntax error. Expected operator"),
-                arguments("@a and or", "Syntax error. Expected operand"),
-                arguments("or or", "Syntax error. Expected operand"),
-                arguments("a b", "Syntax error. Expected operator"),
-                arguments("( a and b ) )", "Syntax error. Unmatched )"),
-                arguments("( ( a and b )", "Syntax error. Unmatched (")
+                arguments("@a @b or", "Tag expression '@a @b or' could not be parsed because of syntax error: expected operator"),
+                arguments("@a and (@b not)", "Tag expression '@a and (@b not)' could not be parsed because of syntax error: expected operator"),
+                arguments("@a and (@b @c) or", "Tag expression '@a and (@b @c) or' could not be parsed because of syntax error: expected operator"),
+                arguments("@a and or", "Tag expression '@a and or' could not be parsed because of syntax error: expected operand"),
+                arguments("or or", "Tag expression 'or or' could not be parsed because of syntax error: expected operand"),
+                arguments("a b", "Tag expression 'a b' could not be parsed because of syntax error: expected operator"),
+                arguments("( a and b ) )", "Tag expression '( a and b ) )' could not be parsed because of syntax error: unmatched )"),
+                arguments("( ( a and b )", "Tag expression '( ( a and b )' could not be parsed because of syntax error: unmatched (")
         );
     }
 
     @ParameterizedTest
     @MethodSource("data")
-    public void parser_expression(final String infix, final String expectedError) {
-        try {
-            parser.parse(infix);
-            fail();
-        } catch (TagExpressionException e) {
+    public void parser_expression(String infix, String expectedError) {
+	TagExpressionException e = assertThrows(TagExpressionException.class,
+			() -> TagExpressionParser.parse(infix));
+
             assertEquals(expectedError, e.getMessage());
-        }
     }
 
 }