+++ messages/CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-* Add `id` field to `Background`, `Rule` and `Examples`. Needed for search engine indexing.
-
 ### Changed
 
 ### Deprecated
@@ -35,7 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   * JavaScript: `import { version } from '@cucumber/messages'`
 * Add `Meta` message
   ([#976](https://github.com/cucumber/cucumber/pull/976)
-   [aslakhellesoy]) 
+   [aslakhellesoy])
 * [Java] Upgrade to `cucumber-parent:2.1.0` (needed to expose library version)
  ([#976](https://github.com/cucumber/cucumber/pull/976)
   [aslakhellesoy])